+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Features
 - Change field type to password for change password form (147270697)
+- Added `Merchant Liable` concept to Installation and Application (147092245)
 
 # 1.19.1
 2017-05-12